+++ engineFoxtrot/Engine/SceneManager/SpriteObject.cpp
@@ -2,7 +2,7 @@
 #include "SpriteObject.h"
 
 SpriteObject::SpriteObject() {}
-SpriteObject::SpriteObject(const int _textureID, const int _height, const int _width, const int _size, int _animationDelay, const char* _fileName) :
+SpriteObject::SpriteObject(const int _textureID, const float _height, const float _width, const int _size, int _animationDelay, const char* _fileName) :
 							textureID(_textureID), height(_height), width(_width), lengthOfTextures(_size), animationDelay(_animationDelay), fileName(_fileName){};
 SpriteObject::~SpriteObject() {};
 
@@ -22,7 +22,7 @@ void SpriteObject::setAnimationDelay(int delay) { animationDelay = delay; }
 /// Identifier of the tick of a frame. 
 /// @return 
 /// a left poition of full texture
-int SpriteObject::getLeftPos(uint32_t tick){
+float SpriteObject::getLeftPos(uint32_t tick){
 	uint32_t pos = tick % lengthOfTextures;
 	return pos * width;
 }
